added seperate file for config parsing
