+++ src/workers/worker_tile_processor.py
@@ -2,9 +2,9 @@ import os.path
 import shutil
 import time
 import rioxarray
-from city_data import CityData, unpack_quoted_value
+from city_data import CityData
 from src.constants import FILENAME_ERA5, METHOD_TRIGGER_ERA5_DOWNLOAD, PROCESSING_METHODS
-from src.workers.worker_tools import create_folder
+from src.workers.worker_tools import create_folder, unpack_quoted_value
 from worker_tools import reverse_y_dimension_as_needed, save_tiff_file
 
 PROCESSING_PAUSE_TIME_SEC = 30
